Add 4 new recipes
This commit adds 4 new recipes to the repository: "Milk Ramen", "Peri Peri Chicken Drumsticks", "Mac n Cheese", and "Grilled Sub Sandwich".

These recipes provide a diverse range of options for anyone looking to try out something new in the kitchen. The recipes are added to the "recipes" folder in the repository.

Thanks for your attention to this update!